requirements.txt
autopep8==1.5.7
Django==2.2.10
django-cors-headers==3.8.0
django-filter==21.1
djangorestframework==3.12.4
flake8==3.9.2
Markdown==3.3.4
mccabe==0.6.1
pycodestyle==2.7.0
pyflakes==2.3.1
python-dotenv==0.19.0
pytz==2021.1
sqlparse==0.4.2
toml==0.10.2
